+++ tests/calibnet_wallet.rs
@@ -5,7 +5,7 @@
 //! and assumes:
 //! - `forest-wallet` is on `PATH`,
 //! - a Forest daemon is running and synced to calibnet,
-//! - [`PRELOADED_ADDRESS`] is funded and imported into both backends,
+//! - [`FOREST_TEST_PRELOADED_ADDRESS`] is funded and imported into both backends (env var of the same name; see `forest_wallet_init`),
 //! - `FULLNODE_API_INFO` is exported.
 
 #[path = "common/calibnet_wallet_helpers.rs"]
@@ -14,10 +14,6 @@ mod helpers;
 use helpers::*;
 use rstest::rstest;
 use serde_json::json;
-use tokio::sync::OnceCell;
-
-/// Funded delegated wallet shared across the delegated tests.
-static FUNDED_DELEGATED: OnceCell<String> = OnceCell::const_new();
 
 #[rstest]
 #[case::local(Backend::Local)]
@@ -51,8 +47,8 @@ async fn market_add_balance_message_on_chain() {
     let result = rpc_call(
         "Filecoin.MarketAddBalance",
         json!([
-            PRELOADED_ADDRESS.as_str(),
-            PRELOADED_ADDRESS.as_str(),
+            FOREST_TEST_PRELOADED_ADDRESS.as_str(),
+            FOREST_TEST_PRELOADED_ADDRESS.as_str(),
             ATTO_FIL,
         ]),
     )
@@ -69,7 +65,7 @@ async fn market_add_balance_message_on_chain() {
 #[ignore = "requires a running calibnet Forest daemon"]
 async fn send_to_filecoin_address(#[case] backend: Backend) {
     let target = wallet(backend, &["new"]).unwrap();
-    let msg = send_from(&PRELOADED_ADDRESS, &target, FIL_AMT, backend).unwrap();
+    let msg = send_from(&FOREST_TEST_PRELOADED_ADDRESS, &target, FIL_AMT, backend).unwrap();
     eprintln!("send to {target} ({}) msg: {msg}", backend.label());
     let funded = poll_until_funded(&target, backend).await.unwrap();
     eprintln!("{target} funded balance: {funded}");
@@ -82,7 +78,7 @@ async fn send_to_filecoin_address(#[case] backend: Backend) {
 #[ignore = "requires a running calibnet Forest daemon"]
 async fn send_to_eth_equivalent(#[case] backend: Backend) {
     let target = wallet(backend, &["new"]).unwrap();
-    let initial_msg = send_from(&PRELOADED_ADDRESS, &target, FIL_AMT, backend).unwrap();
+    let initial_msg = send_from(&FOREST_TEST_PRELOADED_ADDRESS, &target, FIL_AMT, backend).unwrap();
     eprintln!(
         "initial send to {target} ({}) msg: {initial_msg}",
         backend.label(),
@@ -90,7 +86,7 @@ async fn send_to_eth_equivalent(#[case] backend: Backend) {
     let baseline = poll_until_funded(&target, backend).await.unwrap();
 
     let eth = filecoin_to_eth(&target).await.unwrap();
-    let eth_msg = send_from(&PRELOADED_ADDRESS, &eth, FIL_AMT, backend).unwrap();
+    let eth_msg = send_from(&FOREST_TEST_PRELOADED_ADDRESS, &eth, FIL_AMT, backend).unwrap();
     eprintln!("send to ETH {eth} (mapped from {target}) msg: {eth_msg}");
 
     let updated = poll_until_changed(&target, &baseline, backend)
@@ -165,30 +161,4 @@ async fn delegated_remote_send() {
         observed != baseline,
         "{target} balance unchanged after delegated --remote-wallet send: {observed}",
     );
-}
-
-/// Delegated signer: create once on local, fund locally, mirror to remote
-/// for tests that query or sign.
-async fn funded_delegated_addr() -> &'static str {
-    let addr = FUNDED_DELEGATED
-        .get_or_try_init(|| async {
-            let addr = wallet(Backend::Local, &["new", "delegated"]).unwrap();
-            let fund_msg =
-                send_from(&PRELOADED_ADDRESS, &addr, DELEGATE_FUND_AMT, Backend::Local).unwrap();
-            eprintln!("delegated funding send to {addr} msg: {fund_msg}");
-            let funded = poll_until_funded(&addr, Backend::Local).await.unwrap();
-            eprintln!("delegated wallet {addr} funded balance: {funded}");
-
-            let exported = export_to_temp_file(&addr, Backend::Local).unwrap();
-            let path = exported
-                .path()
-                .to_str()
-                .expect("temp path is not valid UTF-8");
-            let mirrored = wallet(Backend::Remote, &["import", path]).unwrap();
-            assert_eq!(mirrored, addr, "mirror mismatch: {mirrored} != {addr}",);
-            Ok::<_, anyhow::Error>(addr)
-        })
-        .await
-        .unwrap();
-    addr.as_str()
 }
